+++ CHANGES.md
@@ -22,6 +22,7 @@ Changes in 11.5
 - Add "Final Grading Percentages are not configured." warning in InputFinalGrades.php
 - Add ProperTime() function in Date.php
 - License change from GNU/GPLv2 to GNU/GPLv2 or later so it is compatible with GNU/GPLv3 in COPYRIGHT, composer.json & package.json
+- Unescape DB strings (`$_REQUEST` only, `$_GET` is not escaped) in PreparePHP_SELF.fnc.php
 
 Changes in 11.4.4
 -----------------